Create Tetrads Screne Basic Functionality Added
Scrolling is possible, as well as you can click on chromosomes and change colours of centromeres. After clicking, they are associated with one of the boxes on the right and you can navigate to the correct tetrads by simply clicking on that box.

TODO - chromosomes miniature should appear in black box

TODO 2 - magnifying class showing chromosome details